Jar didn't work yet, second try plus rename of SaveData to SaveHandler
