--- pokemonsave.txt
@@ -1,69 +1,0 @@
-<playerData><world>waterpath.tmx</world><coordinates>9,13</coordinates><spawnWorld>tiletown.tmx</spawnWorld><spawnCoord>9,18</spawnCoord><money>0</money><pokemonData><creature><index>1</index><level>5</level><progress>34</progress><health>15.109531772575249</health></creature></pokemonData><items>0,1</items></playerData>
-<worldData>
-<world>
-<name>tiletown.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal0</name><coordinates>10,0</coordinates><destination>waterpath.tmx</destination><destcoord>9,18</destcoord></entity>
-<entity><type>Barrier</type><name>barrier0</name><coordinates>9,19</coordinates></entity>
-<entity><type>Salesman</type><name>salesman0</name><coordinates>7,13</coordinates><inventory>1</inventory></entity>
-<entity><type>Npc</type><name>npc1</name><imageId>0</imageId><coordinates>10,11</coordinates><xRoute>7,12</xRoute><yRoute>11,11</yRoute><currentDestination>1</currentDestination></entity>
-</entities>
-</world>
-<world>
-<name>waterpath.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal1</name><coordinates>10,0</coordinates><destination>watertown.tmx</destination><destcoord>9,18</destcoord></entity>
-<entity><type>Portal</type><name>portal00</name><coordinates>9,19</coordinates><destination>tiletown.tmx</destination><destcoord>10,1</destcoord></entity>
-</entities>
-</world>
-<world>
-<name>grasspath.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal1</name><coordinates>10,0</coordinates><destination>grasstown.tmx</destination><destcoord>9,18</destcoord></entity>
-<entity><type>Portal</type><name>portal3</name><coordinates>9,19</coordinates><destination>watertown.tmx</destination><destcoord>10,1</destcoord></entity>
-</entities>
-</world>
-<world>
-<name>rockpath.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal1</name><coordinates>10,0</coordinates><destination>rocktown.tmx</destination><destcoord>9,18</destcoord></entity>
-<entity><type>Portal</type><name>portal3</name><coordinates>9,19</coordinates><destination>grasstown.tmx</destination><destcoord>10,1</destcoord></entity>
-</entities>
-</world>
-<world>
-<name>watertown.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal2</name><coordinates>9,19</coordinates><destination>waterpath.tmx</destination><destcoord>10,1</destcoord></entity>
-<entity><type>Healer</type><name>healer0</name><coordinates>9,12</coordinates></entity>
-<entity><type>Item</type><name>item2</name><coordinates>10,17</coordinates><index>0</index></entity>
-<entity><type>Item</type><name>item3</name><coordinates>9,17</coordinates><index>1</index></entity>
-<entity><type>Salesman</type><name>salesman1</name><coordinates>8,17</coordinates><inventory>0,1</inventory></entity>
-<entity><type>TrainerNpc</type><name>Trainer0</name><coordinates>10,1</coordinates><creatures><creature><index>3</index><level>10</level></creature></creatures></entity>
-<entity><type>Portal</type><name>portal4</name><coordinates>10,0</coordinates><destination>grasspath.tmx</destination><destcoord>9,18</destcoord></entity>
-</entities>
-</world>
-<world>
-<name>grasstown.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal2</name><coordinates>9,19</coordinates><destination>grasspath.tmx</destination><destcoord>10,1</destcoord></entity>
-<entity><type>Healer</type><name>healer1</name><coordinates>9,12</coordinates></entity>
-<entity><type>Item</type><name>item3</name><coordinates>10,17</coordinates><index>0</index></entity>
-<entity><type>Item</type><name>item4</name><coordinates>9,17</coordinates><index>1</index></entity>
-<entity><type>Salesman</type><name>salesman2</name><coordinates>8,17</coordinates><inventory>0,1</inventory></entity>
-<entity><type>TrainerNpc</type><name>Trainer1</name><coordinates>10,1</coordinates><creatures><creature><index>1</index><level>13</level></creature></creatures></entity>
-<entity><type>Portal</type><name>portal4</name><coordinates>10,0</coordinates><destination>rockpath.tmx</destination><destcoord>9,18</destcoord></entity>
-</entities>
-</world>
-<world>
-<name>rocktown.tmx</name>
-<entities>
-<entity><type>Portal</type><name>portal2</name><coordinates>9,19</coordinates><destination>rockpath.tmx</destination><destcoord>10,1</destcoord></entity>
-<entity><type>Healer</type><name>healer2</name><coordinates>9,12</coordinates></entity>
-<entity><type>Item</type><name>item4</name><coordinates>10,17</coordinates><index>0</index></entity>
-<entity><type>Item</type><name>item5</name><coordinates>9,17</coordinates><index>1</index></entity>
-<entity><type>Salesman</type><name>salesman3</name><coordinates>8,17</coordinates><inventory>0,1</inventory></entity>
-<entity><type>TrainerNpc</type><name>Trainer2</name><coordinates>10,1</coordinates><creatures><creature><index>2</index><level>16</level></creature></creatures></entity>
-<entity><type>Barrier</type><name>barrier2</name><coordinates>10,0</coordinates></entity>
-</entities>
-</world>
-</worldData>
