--- src/worlds/minicave.txt
@@ -1,22 +1,0 @@
-20,20
-10,10
-741,744,744,744,742,743,744,744,744,744,744,744,744,742,743,744,744,744,744,750,
-802,1111,1111,1111,803,804,1111,1111,1111,1111,1111,1111,1111,803,804,1111,1111,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,865,866,1111,811,
-802,1111,1111,1111,1111,1111,865,866,1111,1111,1111,1111,1111,1111,1111,1111,926,927,1111,811,
-802,1111,1111,1111,1111,1111,926,927,1111,1111,1111,1111,1111,1111,986,987,988,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1047,1048,1049,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1108,1109,1110,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,811,
-802,1111,1111,865,866,1111,1111,1111,1111,1111,1111,925,1111,1111,1111,1111,1111,1111,1111,811,
-802,1111,925,926,927,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,811,
-802,1111,1111,806,807,807,807,810,925,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,811,
-802,1111,1111,867,868,869,870,871,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,811,
-802,1111,1111,867,929,930,931,871,1111,1111,1111,1111,1053,1054,1053,1111,1111,1111,1111,811,
-802,1111,1111,989,990,991,990,993,1111,1111,1111,1111,1111,1111,1114,1111,1111,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1053,1111,1111,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1114,1111,1111,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1053,1111,1111,1111,1111,811,
-802,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1111,1114,1111,1111,1111,1111,811,
-143,143,143,143,143,143,143,143,143,143,143,143,143,143,143,143,143,143,143,143
